+++ src/metadata/mod.rs
@@ -11,7 +11,7 @@ use self::function::Function;
 ///
 /// The Vyper metadata.
 ///
-#[derive(Debug, Default, serde::Serialize, serde::Deserialize, Clone)]
+#[derive(Debug, Default, Clone, serde::Serialize, serde::Deserialize)]
 pub struct Metadata {
     /// The functions metadata.
     pub function_info: BTreeMap<String, Function>,